requirements.txt
absl-py==2.3.1
affine==2.4.0
altair==5.5.0
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
asgiref==3.11.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229914644/work
cachetools==6.2.4
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560622100/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1754405378105/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1745834476052/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
contourpy==1.3.0
cycler==0.12.1
distlib==0.4.0
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.19.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==25.12.19
fonttools==4.60.2
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841376066/work
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.76.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.14.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302889598/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
libmambapy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_libmambapy_1753777003/work/libmambapy
llvmlite==0.43.0
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1753546343304/work
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.15.0
npm==0.1.1
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.18.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.1.1
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
protobuf==6.33.4
pyarrow==21.0.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588445684/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
referencing==0.36.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rich==14.2.0
rpds-py==0.27.1
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1755625075382/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724493881/work
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sqlparse==0.5.5
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
typing_extensions==4.15.0
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
virtualenv==20.36.1
Werkzeug==3.1.4
wrapt==2.0.1
xgboost==2.1.4
zipp==3.23.0
zstandard==0.23.0
